Feat : home Page 타입 models로 이동
