+++ src/components/common/patientListBox/completedRequest/CompletedpatientListBox.tsx
@@ -1,16 +1,11 @@
 import { Stack, styled } from "@mui/material";
 import MoreHorizRoundedIcon from "@mui/icons-material/MoreHorizRounded";
-import { PatientListBoxType } from "..";
 import { roleColor } from "@utils/homePage";
 import { useState } from "react";
 import { ChatBox } from "@components/home";
 import { ReactComponent as SendIcon } from "@/assets/completedRequests/send.svg";
 import { ReactComponent as CheckIcon } from "@/assets/homeIcons/check.svg";
-
-type StaffListBoxProps = {
-  isAccept: boolean;
-  data: PatientListBoxType;
-};
+import { StaffListBoxProps } from "@models/home";
 
 function CompletedPatientListBox({ isAccept, data }: StaffListBoxProps) {
   const roleColorPick = roleColor(data.role);
