+++ src/models/index.ts
@@ -1,4 +1,4 @@
 export type * from "./user";
 export type * from "./signin";
-export type * from "./ward";
 export type * from "./search";
+export type * from "./home";
