+++ src/pages/user/MainHomePage.tsx
@@ -4,6 +4,7 @@ import PatientBox from "@components/common/patientListBox";
 import { waitPatientmockData } from "@components/home/wordMainMockData";
 import { userState } from "@libraries/recoil";
 import layoutState from "@libraries/recoil/layout";
+import { CSwitchType } from "@models/home";
 import { Box, styled } from "@mui/material";
 import { useEffect, useState } from "react";
 import { useNavigate } from "react-router-dom";
@@ -42,11 +43,7 @@ export default function MainHomePage() {
           <SubTitle>
             <SubTitleLeft>
               <span>내 환자만 보기</span>
-              <CSwitch
-                onChange={(el: React.ChangeEvent<HTMLInputElement>) =>
-                  setMainWaitIsMine(el.target.checked)
-                }
-              />
+              <CSwitch onChange={(el: CSwitchType) => setMainWaitIsMine(el.target.checked)} />
             </SubTitleLeft>
             <SubTitleRight>
               <span>직종</span>
@@ -67,11 +64,7 @@ export default function MainHomePage() {
           <SubTitle>
             <SubTitleLeft>
               <span>환자별로 묶기</span>
-              <CSwitch
-                onChange={(el: React.ChangeEvent<HTMLInputElement>) =>
-                  setMainAcceptIsGroup(el.target.checked)
-                }
-              />
+              <CSwitch onChange={(el: CSwitchType) => setMainAcceptIsGroup(el.target.checked)} />
             </SubTitleLeft>
           </SubTitle>
           {waitPatientmockData.mainAccept.map(el => (
